Restructure as a single file
It can just be '$ scp'd anywhere!!
