+++ pyin.py
@@ -1,0 +1,434 @@
+"""
+Core components for pyin
+"""
+
+
+from __future__ import division
+
+import argparse
+import functools
+import itertools
+import json
+import operator
+import os
+import re
+import traceback
+from typing import Optional, TextIO
+from types import GeneratorType
+
+
+__all__ = ['pmap']
+
+
+__version__ = '0.5.4'
+__author__ = 'Kevin Wurster'
+__license__ = '''
+New BSD License
+
+Copyright (c) 2015-2023, Kevin D. Wurster
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+* Redistributions of source code must retain the above copyright notice, this
+  list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright notice,
+  this list of conditions and the following disclaimer in the documentation
+  and/or other materials provided with the distribution.
+
+* The names of pyin its contributors may not be used to endorse or
+  promote products derived from this software without specific prior written
+  permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+'''
+
+
+def _importer(string, scope):
+
+    """
+    Parse expressions and import modules into a scope.
+    """
+
+    matches = set(re.findall("([a-zA-Z_.][a-zA-Z0-9_.]*)", string))
+    for m in matches:
+
+        split = m.split('.', 1)
+
+        # Like: json.loads(line)
+        if len(split) == 1:
+            module = split[0]
+            other = []
+        # Like: tests.module.upper(line)
+        elif len(split) == 2:
+            module, other = split
+            other = [other]
+        # Shouldn't hit this
+        else:  # pragma no cover
+            raise RuntimeError("Error importing: {}".format(m))
+
+        # Are you trying to figure out why relative imports don't work?  If so,
+        # the issue is probably `m.split()` producing ['', 'name'] instead of
+        # ['.name'].  `__import__('.name')__` doesn't appear to work though,
+        # so good luck!
+        if not module:
+            continue
+
+        try:
+            scope[module] = __import__(
+                module,
+                fromlist=other,
+                level=0)
+        except ImportError:
+            pass
+
+    return scope
+
+
+def pmap(expressions, iterable, var='line'):
+
+    """
+    Like `map()` but with `eval()` and multiple Python expressions.
+    Expressions can access the current line being processed with a variable
+    called `line`.
+
+    It's like being dropped inside of a `for` loop and with the ability to
+    create simple `if` statements and variable re-assignment.  This:
+
+        for line in iterable:
+            if 'mouse' in line:
+                line = line.upper()
+            if 'cat' in line:
+                yield line
+
+    turns into this:
+
+        expressions = (
+            "line.upper() if 'mouse' in var else line",
+            "'cat' in var")
+        pmap(expressions, iterable, var='line')
+
+    Expressions are Python code that are passed through `eval()`, which is
+    given a modified global and local scope to help prevent mishaps by default.
+    External modules are automatically imported so don't pass any code to
+    `pmap()` that you wouldn't pass to `eval()`.
+
+    Expressions can be used for a limited amount of control flow and filtering
+    depending on what comes out of `eval()`.  There are 3 types of objects
+    that an expression can return, and in some cases the object stored in `var`
+    is changed:
+
+        True : The original object from the `iterator` is passed to the next
+               expression for evaluation.  If the last expression produces
+               `True` the object is yielded.  Whatever is currently stored
+               in `var` remains unchanged.
+
+        False or None: Expression evaluation halts and processing moves on to
+                       the next object from `iterator`.
+
+        <object> : Expression results that aren't `True` or `False` are
+                   placed inside of the `var` name for subsequent expressions.
+
+    So, use expressions that evaluate as `True` or `False` to filter lines and
+    anything else to transform content.
+
+    Consider the following example CSV and expressions using `line` as `var`:
+
+        "field1","field2","field3"
+        "l1f1","l1f2","l1f3"
+        "l2f1","l2f2","l3f3"
+        "l3f1","l3f2","l3f3"
+        "l4f1","l4f2","l4f3"
+        "l5f1","l5f2","l5f3"
+
+    This expression would only return the first line because it evaluates as
+    `True` and the rest evaluate as `False`:
+
+        "'field' in line"
+
+        is the same as:
+
+        for line in iterator:
+            if 'field' in line:
+                yield line
+
+    To capitalize the second line:
+
+        "line.upper() if 'l2' in line"
+
+        is the same as:
+
+        for line in iterator:
+           if 'l2' in line:
+               line = line.upper()
+
+    Convert to JSON encoded lists:
+
+        "map(lambda x: x.replace('\"', ''), line.strip().split(','))"
+
+        is the same as (this could be condensed):
+
+        for line in iterator:
+            line = line.strip().replace('"', '').strip()
+            yield line.split(',')
+
+
+    Parameters
+    ----------
+    expressions : str or tuple
+        A single Python expression or tuple containing multiple expressions.
+        Expressions are Python code that can reference external modules, which
+        will automatically be imported.
+    iterable : hasattr('__iter__')
+        An iterator producing one object to be evaluated per iteration.
+    var : str, optional
+        Expressions reference this variable to access objects from `iterator`
+        during processing.
+    """
+
+    if isinstance(expressions, str):
+        expressions = expressions,
+    else:
+        expressions = tuple(expressions)
+
+    global_scope = {
+        'it': itertools,
+        'op': operator,
+        'reduce': functools.reduce}
+
+    for expr in expressions:
+        global_scope.update(_importer(expr, global_scope))
+
+    compiled_expressions = []
+    for expr in expressions:
+        try:
+            compiled_expressions.append(compile(expr, '<string>', 'eval'))
+        except SyntaxError:
+            raise SyntaxError(
+                "Could not compile expression:"
+                " {expression}{ls}{ls}{traceback}".format(
+                    expression=expr,
+                    ls=os.linesep,
+                    traceback=traceback.format_exc(0)))
+
+    for idx, obj in enumerate(iterable):
+
+        for expr in compiled_expressions:
+
+            result = eval(expr, global_scope, {'idx': idx, var: obj})
+
+            # Got a generator.  Expand and continue.
+            if isinstance(result, GeneratorType):
+                obj = list(result)
+
+            # Result is some object.  Pass it back in under `var`.
+            elif not isinstance(result, bool):
+                obj = result
+
+            # Result is True/False.  Only continue if True.
+            # Have to explicitly let string_types through for ''
+            # which would otherwise be ignored.
+            elif isinstance(result, str) or result:
+                continue
+
+            # Got something else?  Halt processing for this obj.
+            else:
+                obj = None
+                break
+
+        # Have to explicitly let string_types through for ''
+        # which would otherwise be ignored.
+        if isinstance(obj, str) or obj:
+            yield obj
+
+
+###############################################################################
+# Command Line Interface
+
+
+_CLI_DESCRIPTION = r"""
+It's like sed, but Python!
+
+Map Python expressions across lines of text.  If an expression evaluates as
+'False' or 'None' then the current line is thrown away.  If an expression
+evaluates as 'True' then the next expression is evaluated.  If a list or
+dictionary is encountered it is JSON encoded.  All other objects are cast
+to string.
+
+Newline characters are stripped from the end of each line before processing
+and are added on write unless disabled with '--no-newline'.
+
+This utility employs 'eval()' internally but uses a limited scope to help
+prevent accidental side effects, but there are plenty of ways to get around
+this so don't pass anything through pyin that you wouldn't pass through
+'eval()'.
+
+Remove lines that do not contain a specific word:
+
+    $ pyin -i file "'word' in line"
+
+Capitalize lines containing a specific word:
+
+    $ pyin -i file "line.upper() if 'word' in line else line"
+
+Only print every other word from lines that contain a specific word:
+
+    $ pyin \
+      -i file \
+      "'word' in line" \
+      "line.split()[::2])" \
+      "' '.join(line)"
+
+The '$ pyin' expressions in the command above do:
+
+    1. Select only lines containing 'word'.
+    2. Split string on whitespace and select every other word. Note that this
+       expression results in a list.
+    3. Join list of words into a string.
+""".strip()
+
+
+def cli_parser() -> argparse.ArgumentParser:
+
+    """Construct an :obj:`argparse.ArgumentParser`.
+
+    Provided as an entrypoint to argument parsing that can provide a better
+    entrypoint to :func:`main` from Python.
+    """
+
+    aparser = argparse.ArgumentParser(
+        description=_CLI_DESCRIPTION,
+        formatter_class=argparse.RawDescriptionHelpFormatter
+    )
+
+    aparser.add_argument(
+        '--version', action='version', version=__version__
+    )
+    aparser.add_argument(
+        '-i', '--infile', dest='infiles', metavar='PATH',
+        type=argparse.FileType('r'), default=[argparse.FileType('r')('-')],
+        action='append',
+        help="Input text file. Use '-' for stdin. Can be used multiple times"
+             " to specify multiple input files, which is like: $ cat * | pyin"
+    )
+    aparser.add_argument(
+        '-o', '--outfile', metavar='PATH',
+        type=argparse.FileType('w'), default='-',
+        help="Output text file. Use '-' for stdout."
+    )
+    aparser.add_argument(
+        '--block', action='store_true',
+        help="Place all input text into the `line` variable."
+    )
+    aparser.add_argument(
+        '--no-newline', action='store_true',
+        help="Don't ensure each line ends with a newline character."
+    )
+    aparser.add_argument(
+        '--skip', dest='skip_lines', type=int, default=0,
+        help="Skip N lines in the input text stream before processing. When "
+             "operating in block processing mode the lines are skipped before"
+             " the text is converted to a block. When operating on multiple"
+             " input only lines in the first file are skipped."
+    )
+    aparser.add_argument('expressions', nargs='*')
+
+    return aparser
+
+
+def main(
+        infiles: list[TextIO],
+        outfile: TextIO,
+        expressions: list[str],
+        no_newline: bool,
+        block: bool,
+        skip_lines: int
+) -> int:
+    """Command line interface.
+
+    Direct access to the CLI logic. :obj:`argparse.ArgumentParser` can be
+    accessed with :func:`parser`.
+
+    :param infiles:
+        List of input files to read from. Files are concatenated before
+        processing.
+    :param outfile:
+        Write output to this file.
+    :param expressions:
+        Evaluate these ``pyin`` expressions on each line of text from
+        ``infiles``.
+    :param no_newline:
+        Do not append a line separator to the end of each line.
+    :param block:
+        Treat all input lines as a single line of text. Equivalent to reading
+        all input data into a single :obj:`str` and running that through
+        ``expressions``.
+    :param skip_lines:
+        Skip the first N lines. Applied to ``infiles`` _after_ they have
+        been concatenated. Does not skip the first N lines of each file
+        independently.
+
+    :returns:
+        Exit code.
+    """
+
+    input_stream = itertools.chain.from_iterable(infiles)
+
+    for _ in range(skip_lines):
+        try:
+            next(input_stream)
+        except StopIteration:
+            return 0
+
+    if block:
+        iterable = [os.linesep.join((f.read() for f in infiles))]
+    else:
+        iterable = (l.rstrip(os.linesep) for l in input_stream)
+
+    for line in pmap(expressions, iterable):
+
+        if isinstance(line, str):
+            pass
+        elif isinstance(line, (list, tuple, dict)):
+            line = json.dumps(line)
+        else:
+            line = repr(line)
+
+        if not no_newline and not line.endswith(os.linesep):
+            line += os.linesep
+
+        outfile.write(line)
+
+    return 0
+
+
+def _cli_entrypoint(rawargs: Optional[list] = None):
+    """Shim for CLI entrypoint.
+
+    :func:`main` and :func:`parser` provide an entrypoint to the CLI that can
+    also be invoked from Python, but does not work well with ``setuptools``'s
+    entrypoint machinery. This shim provides a bridge.
+
+    :param rawargs:
+        Like :obj:`sys.argv` (used by default) but without the interpreter
+        path. Used in testing.
+    """
+
+    args = cli_parser().parse_args(args=rawargs)
+
+    exit(main(**vars(args)))
+
+
+if __name__ == '__main__':
+    _cli_entrypoint()
