+++ tests/test_core.py
@@ -1,25 +1,21 @@
-"""
-Unittests for: pyin.core
-"""
+"""Tests for core :module:`pyin` functionality."""
 
-
-import sys
 
 import pytest
 
-import pyin.core
+import pyin
 import tests._test_module
 
 
 def test_single_expr():
-    result = list(pyin.core.pmap("20 <= line <= 80", range(100)))
+    result = list(pyin.pmap("20 <= line <= 80", range(100)))
     assert len(result) == len(list(range(20, 81)))
     for item in result:
         assert 20 <= item <= 80
 
 
 def test_with_map():
-    result = list(pyin.core.pmap(
+    result = list(pyin.pmap(
         "list(map(int, line.split('-')))", ['2015-01-01']))
     assert result == [[2015, 1, 1]]
 
@@ -29,12 +25,12 @@ def test_with_map():
 ])
 def test_scope(obj):
     """Make sure specific objects aren't removed from the scope."""
-    for res in pyin.core.pmap(obj, 'word'):
+    for res in pyin.pmap(obj, 'word'):
         pass
 
 
 def test_floating_point_division():
-    result = next(pyin.core.pmap('5 / 3', ['']))
+    result = next(pyin.pmap('5 / 3', ['']))
     assert isinstance(result, float)
     assert 1 < result < 2
 
@@ -45,4 +41,4 @@ def test_floating_point_division():
 ])
 def test_invalid_expression(expression):
     with pytest.raises(SyntaxError):
-        next(pyin.core.pmap(expression, []))
+        next(pyin.pmap(expression, []))
